Appearance and General tabs split up
