+++ shop/src/pages/admin/settings/Settings.js
@@ -11,6 +11,7 @@ import Deployments from './deployments/List'
 import Shipping from './shipping/Edit'
 import Advanced from './Advanced'
 import Checkout from './Checkout'
+import Appearance from './Appearance/Edit'
 import List from './List'
 
 const AdminSettings = () => (
@@ -41,6 +42,9 @@ const AdminSettings = () => (
     </Route>
     <Route path="/admin/settings/checkout">
       <Checkout />
+    </Route>
+    <Route path="/admin/settings/appearance">
+      <Appearance />
     </Route>
     <Route path="/admin/settings/general">
       <General />
