+++ shop/src/pages/admin/settings/payments/List.js
@@ -4,7 +4,6 @@ import pickBy from 'lodash/pickBy'
 import uniqBy from 'lodash/uniqBy'
 import get from 'lodash/get'
 
-import { AllCurrencies } from 'data/Currencies'
 import useShopConfig from 'utils/useShopConfig'
 import useSetState from 'utils/useSetState'
 import useConfig from 'utils/useConfig'
@@ -228,36 +227,6 @@ const PaymentSettings = () => {
         <div className="actions">{actions}</div>
       </h3>
       <div className="shop-settings processors-list">
-        <div className="select-currency">
-          <h4>
-            <fbt desc="admin.settings.payments.storeCurrency">
-              Store currency
-            </fbt>
-          </h4>
-          <div>
-            <div className="description">
-              <fbt desc="admin.settings.payments.storeCurrencyDesc">
-                You should review any potential legal and tax considerations
-                involved with selling in a currency that is different from the
-                one associated with the country your store is located in.
-              </fbt>
-            </div>
-            <select
-              className="form-control"
-              value={state.currency}
-              onChange={(e) =>
-                setState({ hasChanges: true, currency: e.target.value })
-              }
-            >
-              {AllCurrencies.map((currency) => (
-                <option key={currency[0]} value={currency[0]}>
-                  {currency[1]}
-                </option>
-              ))}
-            </select>
-          </div>
-        </div>
-
         <h4>
           <fbt desc="Integrations">Integrations</fbt>
         </h4>
@@ -358,19 +327,3 @@ const PaymentSettings = () => {
 }
 
 export default PaymentSettings
-
-require('react-styl')(`
-  .shop-settings
-    .select-currency
-      margin-top: 1.5rem
-      padding-bottom: 2.5rem
-      border-bottom: 1px solid #cdd7e0
-      margin-bottom: 2rem
-      line-height: normal
-      > div
-        color: #8293a4
-        max-width: 530px
-        .description
-          font-size: 14px
-          margin-bottom: 1rem
-`)
